+++ src/pytgpt/bard/__init__.py
@@ -1,0 +1,5 @@
+from .main import BARD
+
+__info__ = "Interact with Google's Bard"
+
+__all__ = ["BARD"]